安全告警

+++ frameworks/core/components_ng/manager/drag_drop/drag_drop_func_wrapper.cpp
@@ -137,7 +137,7 @@ void EnvelopedDragData(
     CHECK_NULL_VOID(container);
     auto displayInfo = container->GetDisplayInfo();
     CHECK_NULL_VOID(displayInfo);
-    dragAction->displayId = displayInfo->GetDisplayId();
+    dragAction->displayId = static_cast<int32_t>(displayInfo->GetDisplayId());
 
     std::vector<ShadowInfoCore> shadowInfos;
     GetShadowInfoArray(dragAction, shadowInfos);
